requirements.txt
alembic==1.0.8
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
blinker==1.4
bokeh==1.2.0
boto==2.48.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
conda==4.5.4
conda-build==3.10.5
conda-verify==2.0.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.0
Django==2.1.7
docutils==0.14
dominate==2.3.5
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==0.8.17
fastcache==1.0.2
filelock==3.0.4
flake8==3.5.0
Flask==1.0.3
Flask-Admin==1.5.1
Flask-Assets==0.12
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
Flask-Cors==3.0.4
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
flask-mongoengine==0.9.5
Flask-PageDown==0.2.2
Flask-RQ==0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WhooshAlchemy-Redux==0.7.1
Flask-WhooshAlchemyPlus==0.7.5
flask-whooshee==0.6.0
Flask-WTF==0.14.2
future==0.17.1
gcloud==0.18.3
gevent==1.3.0
glob2==0.6
googleapis-common-protos==1.6.0
greenlet==0.4.13
gunicorn==19.9.0
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httpie==1.0.2
httplib2==0.13.1
hyperlink==19.0.0
idna==2.6
imageio==2.3.0
imagesize==1.0.0
incremental==17.5.0
inflection==0.3.1
ipykernel==4.8.2
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.12.0
Jinja2==2.10.1
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
mongoengine==0.15.3
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
mysql-connector==2.1.4
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.5.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.16.4
numpydoc==0.8.0
oauth2client==4.1.3
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
pathspec==0.5.5
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==6.1.0
pipenv==2018.11.26
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
prompt-toolkit==2.0.9
protobuf==3.9.0
psutil==5.4.5
psycopg2==2.7.6.1
psycopg2-binary==2.8.2
py==1.5.3
pyasn1==0.4.3
pyasn1-modules==0.2.5
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
PyDispatcher==2.0.5
pyflakes==1.6.0
pygal==2.4.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==1.8.4
pymongo==3.7.0
PyMySQL==0.9.3
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.4.0
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-cov==2.5.1
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
python-editor==1.0.3
python-slugify==1.2.5
pytz==2019.1
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
Quandl==3.3.0
queuelib==1.5.0
raygun4py==3.1.6
redis==2.10.6
requests==2.22.0
rope==0.10.7
rq==0.12.0
rsa==3.4.2
ruamel-yaml==0.15.35
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.7.3
seaborn==0.9.0
semantic-version==2.5.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sodapy==1.5.4
sortedcollections==0.6.1
sortedcontainers==1.5.10
soupsieve==1.9.2
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
static3==0.7.0
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tabulate==0.7.5
tblib==1.3.2
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
Twisted==19.7.0
typing==3.6.4
unicodecsv==0.14.1
Unidecode==1.0.22
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.5.3
visitor==0.1.3
w3lib==1.21.0
wcwidth==0.1.7
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.15.4
Whoosh==2.6.0
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
xml2json==1.1
xmltodict==0.12.0
zict==0.1.3
zope.interface==4.5.0
